requirements.txt

mediapipe
tensorflow
transformers
matplotlib
tqdm
imutils
dlib
Pillow


fastapi
uvicorn
numpy
opencv-python
scikit-learn
joblib
dlib
imutils
